Ability to produce & consume from service network
Enhance Kafka client properties to be multi-cluster. (Not needed for Streams, which is single cluster, but `KafkaTopicDescriptor` supports `cluster()`, so...)

Wire in remaining properties to allow a streams based microservice to be started and run against local Kafka.  Add test to prove test service can consume and produce to Kafka.