Switch to shared binary clients (#135)
* Switch to shared binary clients

Having per-topic consumers and producers would encourage bad usage patterns, i.e. separate consumers and producers for each topic, where a much cleaner and more efficient approach is a single consumer and producer per app.

For now, just expose shared binary producer and consumers and provide helpers for serde of key and value. It's not particularly pretty, but it'll work and maintains the use of the default Kafka clients.