Basic coercer
part of https://github.com/creek-service/creek-kafka/issues/125

Topic records read from the YAML test files may not have the correct key and value types, e.g. the topic may be have `long` key, but a small number will be deserialized as an int.

Before producing the record to Kafka the types will need to be coerced to match the topic.