System test functional test
part of https://github.com/creek-service/creek-kafka/issues/125

Add the start of a functional test that runs the system test executor for end-to-end testing of the test extension.

This uncovered a bug that the bootstrap servers were not being set. Resolved this by passing the bootstrap config via a properties override provider. This necessitated switching to lazy resolution of property overrides.